Create jobs for devices for notifications
So we can keep track if a particular device failed in sending notification
