+++ app/jobs/announcement_sender_job.rb
@@ -1,0 +1,7 @@
+class AnnouncementSenderJob < ActiveJob::Base
+  queue_as :default
+
+  def perform(announcement, device)
+    AnnouncementSender.new(announcement, device.token).run
+  end
+end
